requirements.txt
grpcio-tools ~= 1.30
pylsl ~= 1.10.4